Added some docs and modified some expected functions
