+++ kotlinpoet/src/commonMain/kotlin/com/squareup/kotlinpoet/jvm/alias/JvmDeclaredType.kt
@@ -16,6 +16,7 @@
 package com.squareup.kotlinpoet.jvm.alias
 
 /**
+ * An expected typealias for `javax.lang.model.type.DeclaredType`.
  *
  * @author ForteScarlet
  */
