+++ kotlinpoet/src/commonMain/kotlin/com/squareup/kotlinpoet/jvm/alias/JvmFile.kt
@@ -16,6 +16,7 @@
 package com.squareup.kotlinpoet.jvm.alias
 
 /**
+ * An expected typealias for `java.io.File`.
  *
  * @author ForteScarlet
  */
