+++ kotlinpoet/src/commonMain/kotlin/com/squareup/kotlinpoet/jvm/alias/JvmModifier.kt
@@ -16,6 +16,7 @@
 package com.squareup.kotlinpoet.jvm.alias
 
 /**
+ * An expected typealias for `javax.lang.model.element.Modifier`.
  *
  * @author ForteScarlet
  */
