+++ kotlinpoet/src/commonMain/kotlin/com/squareup/kotlinpoet/jvm/alias/JvmTypes.kt
@@ -16,7 +16,7 @@
 package com.squareup.kotlinpoet.jvm.alias
 
 /**
- *
+ * An expected typealias for `javax.lang.model.util.Types`
  * @author ForteScarlet
  */
 public expect interface JvmTypes
