--- kotlinpoet/src/commonTest/kotlin/com/squareup/kotlinpoet/CommonClassNameTest.kt
@@ -1,30 +1,0 @@
-/*
- * Copyright (C) 2014 Square, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.squareup.kotlinpoet
-
-import kotlin.test.Test
-import kotlin.test.assertEquals
-
-class CommonClassNameTest {
-
-  @Test
-  fun classNameFromClass() {
-    assertEquals(
-      "kotlin.Any",
-      Any::class.asClassName().toString(),
-    )
-  }
-}
