+++ kotlinpoet/src/jvmMain/kotlin/com/squareup/kotlinpoet/jvm/alias/JvmTypes.jvm.kt
@@ -17,8 +17,5 @@ package com.squareup.kotlinpoet.jvm.alias
 
 import javax.lang.model.util.Types
 
-/**
- *
- * @author ForteScarlet
- */
+
 public actual typealias JvmTypes = Types
